requirements.txt
boto3==1.34.98
botocore==1.34.98
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
idna==3.7
jmespath==1.0.1
jwt==1.3.1
mysql-connector-python==8.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
